stochastic updaates

+++ src/algorithms/em.jl
@@ -921,7 +921,7 @@ function stochastic_m_step!(
 
     # Stochastic updating weight
 
-    ρ = stochastic_config.weight_fn(t)
+    ρ = stochastic_config.weight_fn(t)::Float64
